Extend functions.py to support Python >=2.7
(c) Copyright IBM Corp. 2015
